requirements.txt
pandas
xlrd >= 1.0.0
transformers
umap-learn
